Repeat copy task is implemented and converging, but not generalizing.
It learns to do the task perfectly for all seq lengths and repeat
numbers that have been tried, but chokes as soon as numbers increased
beyond those seen during training. Interesting, as normal copy task
generalizes well. One thing to note is that the read patterns for this
task don't look as nice. That's a possible thing to look into when
trying to fix. Copy task still works despite other changes to code.
On to associative recall task for now.